+++ slick_reporting/form_factory.py
@@ -97,7 +97,7 @@ def _default_foreign_key_widget(f_field):
 
 
 def report_form_factory(model, crosstab_model=None, display_compute_reminder=True, fkeys_filter_func=None,
-                        foreign_key_widget_func=None, excluded_fields=None):
+                        foreign_key_widget_func=None, excluded_fields=None, initial=None, required=None):
     """
     Create a Report Form based on the report_model passed by
     1. adding a start_date and end_date fields
@@ -110,12 +110,16 @@ def report_form_factory(model, crosstab_model=None, display_compute_reminder=Tru
     :param fkeys_filter_func:  a receives an OrderedDict of Foreign Keys names and their model field instances found on the model, return the OrderedDict that would be used
     :param foreign_key_widget_func: receives a Field class return the used widget like this {'form_class': forms.ModelMultipleChoiceField, 'required': False, }
     :param excluded_fields: a list of fields to be excluded from the report form
+    :param initial a dict for fields initial
+    :param required a list of fields that should be marked as required
     :return:
     """
     foreign_key_widget_func = foreign_key_widget_func or _default_foreign_key_widget
     fkeys_filter_func = fkeys_filter_func or (lambda x: x)
 
     # gather foreign keys
+    initial = initial or {}
+    required = required or []
     fkeys_map = get_foreign_keys(model)
     excluded_fields = excluded_fields or []
     for excluded in excluded_fields:
@@ -127,17 +131,21 @@ def report_form_factory(model, crosstab_model=None, display_compute_reminder=Tru
     fields = OrderedDict()
 
     fields['start_date'] = forms.DateTimeField(required=False, label=_('From date'),
-                                               initial=app_settings.SLICK_REPORTING_DEFAULT_START_DATE,
+                                               initial=initial.get('start_date',
+                                                                   app_settings.SLICK_REPORTING_DEFAULT_START_DATE),
                                                widget=forms.DateTimeInput(attrs={'autocomplete': "off"}))
 
     fields['end_date'] = forms.DateTimeField(required=False, label=_('To  date'),
-                                             initial=app_settings.SLICK_REPORTING_DEFAULT_END_DATE,
+                                             initial=initial.get('end_date',
+                                                                 app_settings.SLICK_REPORTING_DEFAULT_END_DATE),
                                              widget=forms.DateTimeInput(attrs={'autocomplete': "off"}))
 
     for name, f_field in fkeys_map.items():
         fkeys_list.append(name)
-
-        fields[name] = f_field.formfield(**foreign_key_widget_func(f_field))
+        field_attrs = foreign_key_widget_func(f_field)
+        if name in required:
+            field_attrs['required'] = True
+        fields[name] = f_field.formfield(**field_attrs)
 
     if crosstab_model and display_compute_reminder:
         fields['crosstab_compute_reminder'] = forms.BooleanField(required=False,