+++ slick_reporting/views.py
@@ -92,7 +92,10 @@ class SlickReportViewBase(FormView):
         """
         return self.form_class or report_form_factory(self.get_report_model(), crosstab_model=self.crosstab_model,
                                                       display_compute_reminder=self.crosstab_compute_reminder,
-                                                      excluded_fields=self.excluded_fields)
+                                                      excluded_fields=self.excluded_fields,
+                                                      initial=self.get_form_initial(),
+                                                      # required=self.required_fields
+                                                      )
 
     def get_form_kwargs(self):
         """
@@ -209,7 +212,8 @@ class SlickReportViewBase(FormView):
     def get_report_slug(cls):
         return cls.__name__.lower()
 
-    def get_initial(self):
+    @staticmethod
+    def get_form_initial():
         # todo revise why not actually displaying datetime on screen
         return {
             'start_date': SLICK_REPORTING_DEFAULT_START_DATE,